mapcast-query/pyproject.toml
[tool.poetry]
name = "mapcast-query"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.34.136"
duckdb = "^1.0.0"
structlog = "^24.2.0"
pyarrow = "^16.1.0"
polars-lts-cpu = "^0.20.31"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
